pyproject.toml
[project]
name = "file-generator-etl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.1"
dependencies = [
    "boto3>=1.35.90",
    "faker>=33.1.0",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "setuptools>=75.6.0",
    "sqlalchemy>=2.0.36",
]
